+++ qiskit/providers/ibmq/job/schema.py
@@ -15,7 +15,7 @@
 """Schemas for job."""
 
 from marshmallow import pre_load
-from marshmallow.validate import Range
+from marshmallow.validate import Range, OneOf
 
 from qiskit.validation import BaseSchema
 from qiskit.validation.fields import Dict, String, Nested, Integer, Boolean, DateTime
@@ -24,6 +24,7 @@ from qiskit.result.models import ResultSchema
 
 from qiskit.providers.ibmq.utils.fields import Enum, map_field_names
 from qiskit.providers.ibmq.apiconstants import ApiJobKind, ApiJobStatus
+from qiskit.providers.ibmq.job.queueinfo import InfoQueueResponseSchema
 
 
 # Mapping between 'API job field': 'IBMQJob attribute', for solving name
@@ -110,3 +111,13 @@ class JobResponseSchema(BaseSchema):
         duplicates the unknown keys.
         """
         return map_field_names(FIELDS_MAP, data)
+
+
+class StatusResponseSchema(BaseSchema):
+    """Schema for StatusResponse"""
+
+    # Optional properties
+    infoQueue = Nested(InfoQueueResponseSchema, required=False)
+
+    # Required properties
+    status = String(required=True, validate=OneOf([status.value for status in ApiJobStatus]))