+++ test/ibmq/test_ibmq_job_attributes.py
@@ -19,7 +19,7 @@ from unittest import mock
 import re
 
 from qiskit import ClassicalRegister, QuantumCircuit, QuantumRegister
-from qiskit.providers import JobStatus
+from qiskit.providers.jobstatus import JobStatus, JOB_FINAL_STATES
 from qiskit.providers.ibmq.job.exceptions import IBMQJobFailureError, JobError
 from qiskit.providers.ibmq.api.clients.account import AccountClient
 from qiskit.providers.ibmq.exceptions import IBMQBackendValueError
@@ -253,29 +253,33 @@ class TestIBMQJobAttributes(JobTestCase):
 
         self.assertEqual(job.batman, 'bruce')
 
-    @requires_provider
-    def test_queue_info(self, provider):
+    # @requires_provider
+    @run_on_device
+    def test_queue_info(self, backend, provider):
         """Test retrieving queue information."""
         # Find the most busy backend.
-        backend = max([b for b in provider.backends(simulator=False) if b.status().operational],
-                      key=lambda b: b.status().pending_jobs)
+        # backend = max([b for b in provider.backends(simulator=False) if b.status().operational],
+        #               key=lambda b: b.status().pending_jobs)
         qobj = assemble(transpile(self._qc, backend=backend), backend=backend)
+        leave_states = list(JOB_FINAL_STATES) + [JobStatus.RUNNING]
         job = backend.run(qobj)
         for _ in range(10):
             status = job.status()
-            if status is JobStatus.QUEUED:
+            # Even if job status is QUEUED, queue information may not be
+            # immediately available.
+            if (status is JobStatus.QUEUED and job.queue_position()) or \
+                    status in leave_states:
                 break
-            time.sleep(1)
+            time.sleep(0.5)
+
         if status is JobStatus.QUEUED:
-            self.assertIsNotNone(
-                job.queue_position(),
-                "Job {} is queued but has no queue position.".format(job.job_id()))
-            self.assertIsNotNone(
-                job.estimated_run_time(),
-                "Job {} is queued but has no estimated time.".format(job.job_id()))
+            msg = "Job {} is queued but has no ".format(job.job_id())
+            self.assertIsNotNone(job.queue_position(), msg + "queue position.")
+            self.assertIsNotNone(job._queue_info, msg + "queue info.")
+            for attr, value in job._queue_info.__dict__.items():
+                self.assertIsNotNone(value, msg + attr)
         else:
             self.assertIsNone(job.queue_position())
-            self.assertIsNone(job.estimated_run_time())
             self.log.warning("Unable to retrieve queue information")
 
         # Cancel job so it doesn't consume more resources.